+++ Library/Instance.cs
@@ -1,6 +1,6 @@
 namespace Library
 {
-	public class Instance(Entity entity, Space space, Coord point, uint instanceFormID, string label, Space? teleportsTo, LockLevel lockLevel, Shape? primitiveShape, float spawnWeight = 1, bool inContainer = false)
+	public class Instance(Entity entity, Space space, Coord point, uint instanceFormID, string label, Space? teleportsTo, LockLevel lockLevel, Shape? primitiveShape, double spawnWeight = 1, bool inContainer = false)
 		: BaseInstance(entity, space, label, lockLevel, spawnWeight, inContainer)
 	{
 		// The coordinates of this instance
